M2: split up ray-tracing into passes
Instead of running ray-intersections for every iteration, run
for a short number of iterations at a time (a 'pass')
before returning to the CPU.

This allows for a smaller output buffer. It also means that the
execution time will be short for each pass, which is important
to stop the shader code from being killed due to timeouts on
non-Mac hardware.