M2: ignore triangles when bouncing from the 'outside'
Problem:
Floating point error means that a ray can occasionally end up outside
the geometry, and bounce off the outside of the triangle it just
intersected with.

Solution:
Perform an orientation pass on the geometry at the start to ensure
that the triangle normals all point outwards. Then when performing
intersections, do not intersect triangles when the dot product with
the ray is negative (i.e. when the ray is approaching from outside).

This is better than just ignoring the last intersected triangle since
there are other edge cases (literally) where a ray could bounce off
the outside of an adjacent triangle.